We have a homepage, a public tracks listing and we can view our friends' tracks now.
Added some info on the homepage.
Added new public tracks listing.
Added the ability to view all tracks (public or not) of my confirmed friends.
Added a Tracks#length method to calculate the total length of the track.
Some changes to the logic & design of the main navigation menu.
Disabled the scrollwheel zooming feature of the Google Maps component (it is annoying when there is a page scroll).
Various other minor changes.